fix tag state not updating correctly
