+++ apps/web/app/environments/[environmentId]/surveys/[surveyId]/responses/ResponseTagsWrapper.tsx
@@ -2,59 +2,18 @@ import TagsCombobox from "@/app/environments/[environmentId]/surveys/[surveyId]/
 import { useResponses } from "@/lib/responses/responses";
 import { removeTagFromResponse, useAddTagToResponse, useCreateTag } from "@/lib/tags/mutateTags";
 import { useTagsForEnvironment } from "@/lib/tags/tags";
-import { cn } from "@formbricks/lib/cn";
-import { XCircleIcon } from "@heroicons/react/24/solid";
 import React, { useEffect, useState } from "react";
 import { toast } from "react-hot-toast";
+import { Tag } from "./Tag";
 
 interface ResponseTagsWrapperProps {
   tags: {
     tagId: string;
     tagName: string;
   }[];
-
   environmentId: string;
   surveyId: string;
   responseId: string;
-}
-
-export function Tag({
-  tagId,
-  tagName,
-  onDelete,
-  tags,
-  setTagsState,
-  highlight,
-}: {
-  tagId: string;
-  tagName: string;
-  onDelete: (tagId: string) => void;
-  tags: ResponseTagsWrapperProps["tags"];
-  setTagsState: (tags: ResponseTagsWrapperProps["tags"]) => void;
-  highlight?: boolean;
-}) {
-  return (
-    <div
-      key={tagId}
-      className={cn(
-        "relative flex items-center justify-between gap-2 rounded-full border bg-slate-600 px-2 py-1 text-slate-100",
-        highlight && "border-2 border-green-600"
-      )}>
-      <div className="flex items-center gap-2">
-        <span className="text-sm">{tagName}</span>
-      </div>
-
-      <span
-        className="cursor-pointer text-sm"
-        onClick={() => {
-          setTagsState(tags.filter((tag) => tag.tagId !== tagId));
-
-          onDelete(tagId);
-        }}>
-        <XCircleIcon fontSize={24} className="h-4 w-4 text-slate-100 hover:text-slate-200" />
-      </span>
-    </div>
-  );
 }
 
 const ResponseTagsWrapper: React.FC<ResponseTagsWrapperProps> = ({
@@ -114,7 +73,7 @@ const ResponseTagsWrapper: React.FC<ResponseTagsWrapperProps> = ({
           searchValue={searchValue}
           setSearchValue={setSearchValue}
           tags={environmentTags?.map((tag) => ({ value: tag.id, label: tag.name })) ?? []}
-          currentTags={tags.map((tag) => ({ value: tag.tagId, label: tag.tagName }))}
+          currentTags={tagsState.map((tag) => ({ value: tag.tagId, label: tag.tagName }))}
           createTag={(tagName) => {
             createTag(
               {
